Rename `trackers.js` to align with command name
