+++ main.js
@@ -46,7 +46,7 @@ const twitterListener = require('./twitterListener.js');
 const alertMaintenance = require('./alertMaintenance.js');
 const deleteMessages = require('./deleteMessages.js');
 const dashboard = require('./dashboard.js');
-const trackers = require('./trackers.js');
+const trackers = require('./tracker.js');
 const openContinents = require('./openContinents.js');
 const auraxiums = require('./auraxiums.js');
 const leaderboard = require('./leaderboard.js');
