+++ unifiedWSHandler.js
@@ -14,7 +14,7 @@ const config = require('./subscriptionConfig.js');
 const territory = require('./territory.js');
 const alerts = require('./static/alerts.json');
 const bases = require('./static/bases.json');
-const trackers = require('./trackers.js');
+const trackers = require('./tracker.js');
 const {serverNames, censusRequest, faction} = require('./utils.js');
 
 const wait = require('util').promisify(setTimeout);
